feat: create CRUD for COmments

+++ server/src/services/delete-post-service.ts
@@ -3,5 +3,9 @@ import { db } from '../db/db'
 import { posts } from '../db/schema'
 
 export async function deletePostService(postId: string) {
+  const [post] = await db.select().from(posts).where(eq(posts.id, postId))
+
+  if (!post) throw new Error('invalid id')
+
   await db.delete(posts).where(eq(posts.id, postId))
 }
